Moved all custom errors into an error module
